Refactor more generic classes
