+++ src/edu/nr/lib/FalconPathPlanner.java
@@ -1,4 +1,4 @@
-package edu.nr.robotics.subsystems.drive;
+package edu.nr.lib;
 
 import java.util.LinkedList;
 import java.util.List;
