+++ src/edu/nr/lib/PID/EncoderPIDSource.java
@@ -1,5 +1,6 @@
-package edu.nr.robotics.subsystems.drive;
+package edu.nr.lib.PID;
 
+import edu.nr.robotics.subsystems.drive.Drive;
 import edu.wpi.first.wpilibj.PIDSource;
 
 public class EncoderPIDSource implements PIDSource
