+++ src/edu/nr/lib/navx/NavX.java
@@ -1,4 +1,4 @@
-package edu.nr.robotics.mxp;
+package edu.nr.lib.navx;
 
 import edu.wpi.first.wpilibj.SerialPort;
