+++ src/edu/nr/robotics/auton/AutonDoNothingCommand.java
@@ -1,6 +1,6 @@
 package edu.nr.robotics.auton;
 
-import edu.nr.robotics.subsystems.CMD;
+import edu.nr.lib.CMD;
 
 public class AutonDoNothingCommand extends CMD
 {
