+++ src/edu/nr/robotics/commandgroup/CancelAllCommand.java
@@ -1,6 +1,6 @@
 package edu.nr.robotics.commandgroup;
 
-import edu.nr.robotics.subsystems.CMD;
+import edu.nr.lib.CMD;
 import edu.nr.robotics.subsystems.binGrabber.BinGrabber;
 import edu.nr.robotics.subsystems.drive.Drive;
 import edu.nr.robotics.subsystems.frontElevator.FrontElevator;
