+++ src/edu/nr/robotics/subsystems/drive/commands/DriveIdleCommand.java
@@ -1,6 +1,6 @@
 package edu.nr.robotics.subsystems.drive.commands;
 
-import edu.nr.robotics.subsystems.CMD;
+import edu.nr.lib.CMD;
 import edu.nr.robotics.subsystems.drive.Drive;
 
 /**
