+++ src/edu/nr/robotics/subsystems/drive/commands/DriveJoystickCommand.java
@@ -1,7 +1,7 @@
 package edu.nr.robotics.subsystems.drive.commands;
 
+import edu.nr.lib.CMD;
 import edu.nr.robotics.OI;
-import edu.nr.robotics.subsystems.CMD;
 import edu.nr.robotics.subsystems.drive.Drive;
 import edu.nr.robotics.subsystems.drive.gyro.AngleGyroCorrection;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
