+++ src/edu/nr/robotics/subsystems/drive/commands/DrivePositionCommand.java
@@ -1,12 +1,12 @@
 package edu.nr.robotics.subsystems.drive.commands;
 
+import edu.nr.lib.CMD;
+import edu.nr.lib.FalconPathPlanner;
+import edu.nr.lib.Position;
+import edu.nr.lib.Setpoint;
 import edu.nr.robotics.FieldCentric;
 import edu.nr.robotics.RobotMap;
-import edu.nr.robotics.subsystems.CMD;
 import edu.nr.robotics.subsystems.drive.Drive;
-import edu.nr.robotics.subsystems.drive.FalconPathPlanner;
-import edu.nr.robotics.subsystems.drive.Position;
-import edu.nr.robotics.subsystems.drive.Setpoint;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 /**
